Switch team schedule view from teams app to matchups app (its model is Matchup so this makes more sense). Modify the view to override queryset and provide context in a cleaner way that requires less duplicate queries.
