+++ apps/matchups/urls.py
@@ -11,6 +11,8 @@ urlpatterns = [
          views.WeeklyMatchupsView.as_view(), name='weekly_matchups_by_week'),
     path('<slug:league>/matchups/<slug:slug>/',
          views.MatchupDetailView.as_view(), name='matchup_detail'),
+    path('<slug:league>/schedule/<slug:team>/',
+         views.TeamScheduleView.as_view(), name='team_schedule'),
     path('<slug:league>/playoffs/',
          views.PlayoffsView.as_view(), name='playoffs'),
 ]
