+++ apps/matchups/views.py
@@ -1,9 +1,13 @@
 import copy
 
+from django.db.models import Q
+from django.http import Http404
+from django.shortcuts import get_object_or_404
 from django.views.generic import DetailView, ListView
 
 from apps.leagues.mixins import LeagueContextMixin
 from apps.leagues.permissions import IsLeagueOwner
+from apps.teams.models import Team
 
 from .models import Matchup
 
@@ -62,6 +66,45 @@ class MatchupDetailView(IsLeagueOwner, LeagueContextMixin, DetailView):
         if league_slug is not None:
             return queryset.with_cases().filter(season__league__slug=league_slug)
         return queryset
+
+
+class TeamScheduleView(IsLeagueOwner, LeagueContextMixin, ListView):
+    """
+    View the schedule of matchups for an individual team's current season.
+    """
+
+    model = Matchup
+    context_object_name = "matchups"
+    template_name = "matchups/team_schedule.html"
+    team = None
+
+    def get_queryset(self):
+        queryset = super().get_queryset()
+
+        # From LeagueContextMixin
+        context = super().get_context_data(object_list=queryset)
+        league = context["league"]
+        season = context["season"]
+
+        team_slug = self.kwargs.get("team")
+        if team_slug is not None:
+            team = get_object_or_404(Team, league=league, slug=team_slug)
+            self.team = team
+        else:
+            raise Http404
+
+        return queryset.filter(
+            Q(home_team=team) | Q(away_team=team), season=season
+        ).order_by("week_number")
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+
+        context["team"] = self.team
+        context["bye_week"] = self.team.bye_week
+        context["teams"] = Team.objects.filter(league=context["league"])
+
+        return context
 
 
 class PlayoffsView(IsLeagueOwner, LeagueContextMixin, ListView):
