+++ templates/leagues/league_detail.html
@@ -62,7 +62,7 @@
         <a href="{% url 'teams:depth_chart' league.slug myteam.team.slug %}" class="btn btn-outline-dark" role="button">
           Depth Chart
         </a>
-        <a href="{% url 'teams:team_schedule' league.slug myteam.team.slug %}" class="btn btn-outline-dark" role="button">
+        <a href="{% url 'matchups:team_schedule' league.slug myteam.team.slug %}" class="btn btn-outline-dark" role="button">
           Schedule
         </a>
       </div>
