+++ templates/teams/team_detail.html
@@ -51,7 +51,7 @@
     <a href="{% url 'teams:depth_chart' league.slug team.slug %}" class="btn btn-outline-dark" role="button">
       Depth Chart
     </a>
-    <a href="{% url 'teams:team_schedule' league.slug team.slug %}" class="btn btn-outline-dark" role="button">
+    <a href="{% url 'matchups:team_schedule' league.slug team.slug %}" class="btn btn-outline-dark" role="button">
       Schedule
     </a>
   </div>
